Adding home dashboard and react native elements
