+++ package.json
@@ -14,6 +14,7 @@
     "react": "16.8.3",
     "react-dom": "^16.8.6",
     "react-native": "https://github.com/expo/react-native/archive/sdk-33.0.0.tar.gz",
+    "react-native-elements": "^1.1.0",
     "react-native-gesture-handler": "^1.2.2",
     "react-native-web": "^0.11.4",
     "react-navigation": "^3.11.0",
